修复：收敛 Rust 1.96 剩余 clippy 基线
采用 checked_div、sort_by_key、Option::zip 与 map values 等标准接口；为 Bevy ECS 独立系统参数保留函数级定点 lint 边界。

Model: gpt-5.6-sol-xhigh

+++ server/src/npc/spawn/ambient_scheduler.rs
@@ -411,7 +411,7 @@ pub fn decide_ambient_check(
     era_spawn_seed: u64,
 ) -> AmbientCheckOutcome {
     let budget = dead_zone_threat_budget(threat_budget(danger_level), zone_kind);
-    if !should_run_interval(now_tick, budget.spawn_interval_ticks as u32) {
+    if !should_run_interval(now_tick, budget.spawn_interval_ticks) {
         return AmbientCheckOutcome::Throttled;
     }
     if counts_against_threat_budget && alive_count >= budget.max_alive {
